refactor(app): thin lib.rs into src/app/ (Phase 1)
Implements Phase 1 of the Rust structure refactor roadmap
(docs/superpowers/plans/2026-04-23-rust-structure-readability-refactor.md).

The crate root is now a 20-line thin module tree plus `pub use
app::run`. All command dispatch and context-resolution helpers moved
verbatim to `src/app/`:

- `src/app/mod.rs` — `run(cli)` dispatch + `remove_data_dir_if_exists`.
- `src/app/context.rs` — `TargetKind`, `classify_target`,
  `find_dst_separator`, `resolve_target_name`,
  `find_saved_workspace_for_cwd`, `resolve_agent_from_context`,
  `resolve_running_container_from_context`, `remember_last_agent`.

Test Preservation Policy (per plan) honored: all 29 tests from the
old `#[cfg(test)] mod tests` block in `lib.rs` moved to their new
homes, verbatim, none deleted or weakened:

- 21 context-helper tests + 2 local test helpers
  (`config_with_workspace`, `fake_runner_with_running_agents`) now
  co-located in `src/app/context.rs` rather than piled into a single
  `app/tests.rs`.
- 8 workspace create/edit tests, which exercise
  `config::AppConfig::create_workspace` / `edit_workspace` through the
  public crate API, moved to a new integration test file
  `tests/workspace_config_crud.rs`. They're kept together for this PR;
  Phase 4 (config split) can re-split if useful.

No logic changes. Visibility tightened where safe: all new items in
`src/app/` default to `pub(crate)` except `TargetKind` / `classify_target`
(kept `pub` matching the original, though grep confirms no external
callers today — candidate for demotion in a follow-up). `run` is
re-exported at the crate root for `jackin::run` callers in
`src/main.rs`.

Verification per COMMITS.md:
- `cargo fmt -- --check` — pass
- `cargo clippy` — zero warnings (main baseline: zero)
- `cargo nextest run` — 409 tests pass, 0 skipped
  (count unchanged from before the move)

Phase 0 note: existing tests already cover every function moved in
Phase 1. Gap-filling tests for later phases (mount-flag combinations,
auth-forward modes, hardline state transitions, agent-filter/query
composition) are deferred to the PR that precedes each phase, per the
plan's "add only high-leverage characterization tests where later
movement is risky" guidance.
